More tests, basic validation of filters, filter views refactoring.
