+++ datahub/search/company/views.py
@@ -1,13 +1,13 @@
-from rest_framework.exceptions import ValidationError
-from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from .models import Company
-from .. import elasticsearch
+from ..views import SearchWithFiltersAPIMixin
 
 
-class SearchCompanyAPIView(APIView):
+class SearchCompanyAPIView(SearchWithFiltersAPIMixin, APIView):
     """Filtered company search view."""
+
+    entity = Company
 
     SORT_BY_FIELDS = (
         'account_manager.name',
@@ -37,12 +37,12 @@ class SearchCompanyAPIView(APIView):
 
     FILTER_FIELDS = (
         'account_manager',
-        'alias',
-        'description',
         'export_to_country',
         'future_interest_country',
-        'name',
         'sector',
+        'registered_address_country',
+        'registered_address_postcode',
+        'registered_address_town',
         'trading_address_country',
         'trading_address_postcode',
         'trading_address_town',
@@ -50,36 +50,12 @@ class SearchCompanyAPIView(APIView):
         'uk_region'
     )
 
-    http_method_names = ('post',)
-
-    def post(self, request, format=None):
-        """Performs filtered company search."""
-        filters = {elasticsearch.remap_filter_id_field(field): request.data[field]
-                   for field in self.FILTER_FIELDS if field in request.data}
-
-        original_query = request.data.get('original_query', '')
-
-        sortby = request.data.get('sortby')
-        if sortby:
-            field = sortby.rsplit(':')[0]
-            if field not in self.SORT_BY_FIELDS:
-                raise ValidationError(f'"sortby" field is not one of {self.SORT_BY_FIELDS}.')
-
-        offset = int(request.query_params.get('offset', 0))
-        limit = int(request.query_params.get('limit', 100))
-
-        results = elasticsearch.get_search_by_entity_query(
-            entity=Company,
-            term=original_query,
-            filters=filters,
-            field_order=sortby,
-            offset=offset,
-            limit=limit,
-        ).execute()
-
-        response = {
-            'count': results.hits.total,
-            'results': [x.to_dict() for x in results.hits],
-        }
-
-        return Response(data=response)
+    REMAP_FIELDS = {
+        'account_manager': 'account_manager.id',
+        'export_to_country': 'export_to_countries.id',
+        'future_interest_country': 'future_interest_countries.id',
+        'sector': 'sector.id',
+        'registered_address_country': 'address_country.id',
+        'trading_address_country': 'trading_address_country.id',
+        'uk_region': 'uk_region.id',
+    }
