+++ datahub/search/contact/views.py
@@ -1,13 +1,13 @@
-from rest_framework.exceptions import ValidationError
-from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from .models import Contact
-from .. import elasticsearch
+from ..views import SearchWithFiltersAPIMixin
 
 
-class SearchContactAPIView(APIView):
+class SearchContactAPIView(SearchWithFiltersAPIMixin, APIView):
     """Filtered contact search view."""
+
+    entity = Contact
 
     SORT_BY_FIELDS = (
         'address_country.name',
@@ -43,36 +43,9 @@ class SearchContactAPIView(APIView):
         'address_country',
     )
 
-    http_method_names = ('post',)
-
-    def post(self, request, format=None):
-        """Performs filtered contact search."""
-        filters = {elasticsearch.remap_filter_id_field(field): request.data[field]
-                   for field in self.FILTER_FIELDS if field in request.data}
-
-        original_query = request.data.get('original_query', '')
-
-        sortby = request.data.get('sortby')
-        if sortby:
-            field = sortby.rsplit(':')[0]
-            if field not in self.SORT_BY_FIELDS:
-                raise ValidationError(f'"sortby" field is not one of {self.SORT_BY_FIELDS}.')
-
-        offset = int(request.data.get('offset', 0))
-        limit = int(request.data.get('limit', 100))
-
-        results = elasticsearch.get_search_by_entity_query(
-            entity=Contact,
-            term=original_query,
-            filters=filters,
-            field_order=sortby,
-            offset=offset,
-            limit=limit,
-        ).execute()
-
-        response = {
-            'count': results.hits.total,
-            'results': [x.to_dict() for x in results.hits],
-        }
-
-        return Response(data=response)
+    REMAP_FIELDS = {
+        'company_name': 'company.name_trigram',
+        'company_sector': 'company_sector.id',
+        'company_uk_region': 'company_uk_region.id',
+        'address_country': 'address_country.id',
+    }
